Move relevant functions to ndpi_util API module
